+++ services/spar/spar.cabal
@@ -51,10 +51,10 @@ library
       Spar.Sem.DefaultSsoCode.Spec
       Spar.Sem.GalleyAccess
       Spar.Sem.GalleyAccess.Http
-      Spar.Sem.IdP
-      Spar.Sem.IdP.Cassandra
-      Spar.Sem.IdP.Mem
-      Spar.Sem.IdP.Spec
+      Spar.Sem.IdPConfigStore
+      Spar.Sem.IdPConfigStore.Cassandra
+      Spar.Sem.IdPConfigStore.Mem
+      Spar.Sem.IdPConfigStore.Spec
       Spar.Sem.IdPRawMetadataStore
       Spar.Sem.IdPRawMetadataStore.Cassandra
       Spar.Sem.IdPRawMetadataStore.Mem
@@ -727,7 +727,7 @@ test-suite spec
       Test.Spar.ScimSpec
       Test.Spar.Sem.DefaultSsoCodeSpec
       Test.Spar.Sem.IdPRawMetadataStoreSpec
-      Test.Spar.Sem.IdPSpec
+      Test.Spar.Sem.IdPConfigStoreSpec
       Test.Spar.Sem.NowSpec
       Test.Spar.Sem.ScimExternalIdStoreSpec
       Test.Spar.TypesSpec
