+++ services/spar/src/Spar/Scim/Auth.hs
@@ -54,8 +54,8 @@ import qualified Spar.Intra.BrigApp as Intra.Brig
 import Spar.Sem.BrigAccess (BrigAccess)
 import qualified Spar.Sem.BrigAccess as BrigAccess
 import Spar.Sem.GalleyAccess (GalleyAccess)
-import Spar.Sem.IdP (IdPConfigStore)
-import qualified Spar.Sem.IdP as IdPConfigStore
+import Spar.Sem.IdPConfigStore (IdPConfigStore)
+import qualified Spar.Sem.IdPConfigStore as IdPConfigStore
 import Spar.Sem.Now (Now)
 import qualified Spar.Sem.Now as Now
 import Spar.Sem.Random (Random)
