+++ services/spar/src/Spar/Sem/IdPConfigStore.hs
@@ -15,7 +15,7 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Spar.Sem.IdP where
+module Spar.Sem.IdPConfigStore where
 
 import Data.Id
 import Imports
